git-svn-id: https://xcmetadataservicestoolkit.googlecode.com/svn/branches/0.3.0@1183 72d1a826-0e72-11de-b5bf-434079a3519f

+++ mst-common/src/java/xc/mst/scheduling/WorkDelegate.java
@@ -28,7 +28,7 @@ public interface WorkDelegate {
 	
 	public int getRecordsProcessed();
 
-	public int getTotalRecords();
+	public long getTotalRecords();
 	
 	public WorkerThread getWorkerThread();
 	public void setWorkerThread(WorkerThread wt);
